Remove duplicate motd pam entries
Due to a bug which is fixed in the previous commit, duplicate entries
have been added to pam.  This fixes that by checking to see if the count
of matching lines is greater than 1.  We assume that we can remove the
last line, as we have only ever added the motd configuration to the last
line of the pam postlogin file.

This is only a cleanup commit and can be removed once the situation has
normalized, meaning that only one pam motd line is present, which should
happen in about 150 puppet runs.